+++ .gitignore
@@ -1,2 +1,3 @@
 snapshots/
 input.json
+__pycache__/
