+++ CHANGELOG.md
@@ -1,33 +1,61 @@
-# CHANGELOG
+# CHANGELOG: Twitter Bird Watcher
+
+## Guiding Principles
+
+* Changelogs are for humans.
+* This changelog attempts to adhere to [Keep a Changlog](https://keepachangelog.com/en/1.0.0/)
+* Our project uses [Semantic Versioning](https://semver.org/)
+* Contributions by community will be credited via their userhandle.
+
+### Types of Changes
+
+* `Added` for new features.
+* `Changed` for changes in existing functionality.
+* `Deprecated` for soon-to-be removed features.
+* `Removed` for now removed features.
+* `Fixed` for any bug fixes.
+* `Security` in case of vulnerabilities.
+
+# Changes
+
+## 0.5.0: Supporting Our Community
+
+* Added modularization of code
+  * Added `tb_watcher.driver_utils`
+  * Added `tb_watcher.swag`
+* Added better logging support via `logging` module
+* Added standardization for CHANGELOG
+* Added better CLI help groups.
+* Fixed README Typos and Divide by Zero Error (@jmallone)
 
 ## 0.4.0: More Algorithms
 
-* Add example README
-* Update README
-* Add scroll algorithm selection
+* Added example README
+* Changed README
+* Added scroll algorithm selection
 
 ## 0.3.0: Remove Ads and Average Scrolling
 
-* Add average scrolling to compensate for scroll heights
-* Add Ad removal logic
-* Fix bug where post length was not respected
+* Added average scrolling to compensate for scroll heights
+* Added Ad removal logic
+* Fixed bug where post length was not respected
 
 ## 0.2.0: Enable Login and New Scroll Method
 
-* Add `--login` for giving users login
-* New aglorithm for scrolling down content to prevent duplications
+* Added `--login` for giving users login
+* Added new aglorithm for scrolling down content to prevent duplications
 
 ## 0.1.1: Better Folder Renames
 
-* Add some code-cleanup
-* Better pop-up removal support
-* Add `--url` support for single profile scraping
-* Add folder renaming to username functionality
-* Add tweet without login limit check
-* Add faster scrolling for centering content
+* Added some code-cleanup
+* Added better pop-up removal support
+* Added `--url` support for single profile scraping
+* Added folder renaming to username functionality
+* Added tweet without login limit check
+* Added faster scrolling for centering content
 
 ## 0.1.0: Initial Release
 
-* Adds support for bio-only
-* Snapshots of Twitter posts
-* Adds max Twitter info
+* Added support for bio-only
+* Added snapshots of Twitter posts
+* Added max Twitter info
