+++ src/tb_watcher/swag.py
@@ -1,0 +1,57 @@
+"""
+Houses the ascii encoded swag.
+By: ProgrammingIncluded
+"""
+
+LOGO = """
+,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,
+,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,
+,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,
+,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,
+,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,'''''''',,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,
+,,,,,,,,,,,,,,,,,,,,,,,,,''',;;:;;:cllollllccc:;,'.',,,,,,,,,,,,,,,,,,,,,,,,,,,,
+,,,,,,,,,,,,,,,,,,,,,,'';:codxkkkkkkkkkkkkkkkkkkxocc:;',,,,,,,,,,,,,,,,,,,,,,,,,
+,,,,,,,,,,,,,,,,,;,'';coxkkkkkkkkkkkkkkkkkkkkkkkkkkkkxl:,',,,,,,,,,,,,,,,,,,,,,,
+,,,,,,,,,,,,,,,,'.,:dkkkkkkkkkkkkkkkkkxocdkkkkkkkkkkxdxxd:,',,,,,,,,,,,,,,,,,,,,
+,,,,,,,,,,,,,'.';lxkkkkkkkkkolxkkkkkkkkl;:okkkkkkkkxxxddxxo;.',,,,,,,,,,,,,,,,,,
+,,,,,,,,'...,:cdxxxkkkkkkkkx;;xkkkkkkkko;cooxkkkkkkkxdocldxdl,',,,,,,,,,,,,,,,,,
+,,,,,,,,,. .',,;:oxkkkkkkkxl::dkkkkkkkko,cOkcdkxxxkkkxl;,;:lol:,,,,,,,,,,,,,,,,,
+,,,,,,,,,,,''..:oxkkkkxdxkocl,:kkkkkxxkl,:ONklokxodkkkxc,;;;:coc,,,,,,,,,,,,,,,,
+,,,,,,,,,,,,,;oxkkkkkkolxx:dk::okkkkxxkl,cONNOccddcokkkl':lcccldc,,,,,,,,,,,,,,,
+,,,,,,,,,,,.:xkxxkkkkxllo:;kkodcoxkkxddc';ccll:..,,'lkkl,:lcclloo,',,,,,,,,,,,,,
+,,,,,,,,,'.,dxxxxkkkkdlxd,':;':c:oxkkc,,.;:,'....;. .cxc,lo:;cloo;.',,,,,,,,,,,,
+,,,,,,,,'.;ooldxxxkkkoll;;c:'.  ,l:cx:.'.;'     'Ox.':;;:oo:;ollol;',,,,,,,,,,,,
+,,,,,,,'..''''cxdxkkko,.lXx.    ;00xo:oKOx'     ;XO'lk:cooo;dNdcoodl,,,;,,,,,,,,
+,,,,,,,,''',,',oc;dkkl..kWo     ;XWWWXNWMNo    .xWOdOocoool;d0oloodxdc,',,,,,,,,
+,,,,,,,,,,,,,,',..;dkl..dWO.    lNMWWWWWWWXl..,dNWWXdcloooc,:lloooodxxdc'.,,,,,,
+,,,,,,,,,,,,,,;,.',,::coONNc   :KWWWWWWWWWWWXKKXNNXdcloooo;,looooooooddxd:.',,,,
+,,,,,,,,,,,,,,,;,.,;'.'kWWWKdokXWWWWNNWWWWWWWWNNNXd:looooc,;clooolcooo::ox:.,;,,
+,,,,,,,,,,,,,,,'.';;;.;0NXXNNWWWWWWWXNWWWWWWWWWXKd:looooc,:c,:doo;.,ll',oo;.,;,,
+,,,,,,,,,,,,''''',;:;..cxk0KXNNNNNXXNNNKOkxollcc,'cooooc,:oc.'odl'.',,.;o;.',,,,
+,,,,,,,,,,,...''.':c;';c;,,:ool:,,,;ldxl,',;,;lc';oool;;:c;...:l,.,;;'',,',,,,,,
+,,,,,,,,,,,,,,,,,',;',::,..,lc'.......;oddkXKd;..;l:,.',,'',;,''.'''...',,,,,,,,
+,,,,,,,,,,,,,,,,,,'...   ..,dxd:,.. .:xkkONWWWO;'clc'.......,;::cll.    .,,,,,,,
+,,,,,,,,,,,,,,,,,,.,ooc'...'O0:'...  ':okOXWWWWl;o;.    ....kOddo0K,    .,,,,,,,
+,,,,,,,,,,,,,,,,'.'xWX0Od, .d:.. ......;dk0kxOO'.'.     ...:xc,;;k0,     ',,,,,,
+,,,,,,,,,,,,,,,. ;xONWXOkl..d:.  ,::,. ,oodd,..    .     .xXX0O00XX;     .;,,,,,
+,,,,,,,,,,,,,,,. ;OOXWXk:'..dk,.  .. .,kXOOO,      .     .kWNWWWWWNc     .,,,,,,
+,,,,,,,,,,,,,,,. .xkOK0k, .'lOkc,'..;oO00xOK,      .      dWWWWWWNWd.    .,,,,,,
+,,,,,,,,,,,,,,,'. ;kkxxko.'xOOOOOOkOOOOOOdOK:. .;;..      :XWWWWNXKx.    .,,,,,,
+,,,,,,,,,,,,,,,,,..;odk00d,'d00KKKK0KKK0OkKKdxxoo, .      .;lccc:;,'....',,,,,,,
+,,,,,,,,,,,,,,,,,,;';oxkxxl..looooooooollll:'..   ''........''',,,;;,,,,,,,,,,,,
+,,,,,,,,,,,,,,,,,,,,,,;,'',.                   ..lXO;,;,;,,,,,,,,,,,,,,,,,,,,,,,
+,,,,,,,,,,,,,,,,,,,,,,,,,,,,......         ...',,;xKo',,,,,,,,,,,,,,,,,,,,,,,,,,
+,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,;;,,,,,''',,,,,,,,,,;;'',,,,,,,,,,,,,,,,,,,,,,,,,
+,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,;,,,,,,,,,,,,,,,,,,,,,,,,,,,,
+,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,
+
+"""
+
+TITLE = """
+__________ ___________ __      __          __           .__                     
+\______   \\__    ___//  \    /  \_____  _/  |_   ____  |  |__    ____ _______  
+ |    |  _/  |    |   \   \/\/   /\__  \ \   __\_/ ___\ |  |  \ _/ __ \\_  __ \ 
+ |    |   \  |    |    \        /  / __ \_|  |  \  \___ |   Y  \\  ___/ |  | \/ 
+ |______  /  |____|     \__/\  /  (____  /|__|   \___  >|___|  / \___  >|__|    
+        \/                   \/        \/            \/      \/      \/         
+"""
